+++ tests/constants.test.ts
@@ -181,6 +181,6 @@ describe("Other constants", () => {
   });
 
   test("VERSION", () => {
-    expect(VERSION).toBe("1.6.0");
+    expect(VERSION).toBe("1.8.0");
   });
 });